vidya_ai_microservice/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
opencv-python==4.10.0.84
ultralytics==8.2.103
google-cloud-vision==3.7.4
imagehash==4.3.1
Pillow==10.4.0
xgboost==2.1.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
mlflow==2.15.1
geopy==2.4.1
